Drop Redis and simply keep jobs in memory
Currently persistence for this little project is a bit overkill and
needing to have a Redis instance running makes it slightly less
convenient to deploy. For now, just use an in-memory data structure and
let everything start afresh on reboot. The structure does grow endlessly
but that's fine for now.